Reduce files and flatten directories
