+++ website/other/bezier-rs-demos/src/types.ts
@@ -76,23 +76,20 @@ export type SubpathDemoArgs = {
 	closed: boolean;
 } & DemoArgs;
 
-export type Demo = {
-	inputOptions: InputOption[];
-	sliderData: Record<string, number>;
-	sliderUnits: Record<string, string | string[]>;
-
-	drawDemo(figure: HTMLElement, mouseLocation?: [number, number]): void;
-	onMouseDown(e: MouseEvent): void;
-	onMouseUp(): void;
-	onMouseMove(e: MouseEvent): void;
-	getSliderUnit(variable: string): string;
-};
-
 export const BEZIER_T_VALUE_VARIANTS = ["Parametric", "Euclidean"] as const;
 export const SUBPATH_T_VALUE_VARIANTS = ["GlobalParametric", "GlobalEuclidean"] as const;
 
 export const CAP_VARIANTS = ["Butt", "Round", "Square"] as const;
 export const JOIN_VARIANTS = ["Bevel", "Miter", "Round"] as const;
+
+export const POINT_INDEX_TO_MANIPULATOR: WasmSubpathManipulatorKey[] = ["set_anchor", "set_in_handle", "set_out_handle"];
+
+// Given the number of points in the curve, map the index of a point to the correct manipulator key
+export const MANIPULATOR_KEYS_FROM_BEZIER_TYPE: { [key in BezierCurveType]: WasmBezierManipulatorKey[] } = {
+	Linear: ["set_start", "set_end"],
+	Quadratic: ["set_start", "set_handle_start", "set_end"],
+	Cubic: ["set_start", "set_handle_start", "set_handle_end", "set_end"],
+};
 
 export function getBezierDemoPointDefaults() {
 	// We use a function to generate a new object each time it is called
@@ -115,3 +112,106 @@ export function getBezierDemoPointDefaults() {
 		],
 	};
 }
+
+export function getSubpathDemoArgs(): SubpathDemoArgs[] {
+	// We use a function to generate a new object each time it is called
+	// to prevent one instance from being shared and modified across demos
+	return [
+		{
+			title: "Open Subpath",
+			triples: [
+				[[45, 20], undefined, [35, 90]],
+				[[175, 40], [85, 40], undefined],
+				[[200, 175], undefined, undefined],
+				[[125, 100], [65, 120], undefined],
+			],
+			closed: false,
+		},
+		{
+			title: "Closed Subpath",
+			triples: [
+				[[60, 125], undefined, [65, 40]],
+				[[155, 30], [145, 120], undefined],
+				[
+					[170, 150],
+					[200, 90],
+					[95, 185],
+				],
+			],
+			closed: true,
+		},
+	];
+}
+
+export const tSliderOptions = {
+	variable: "t",
+	inputType: "slider",
+	min: 0,
+	max: 1,
+	step: 0.01,
+	default: 0.5,
+};
+
+export const errorOptions = {
+	variable: "error",
+	inputType: "slider",
+	min: 0.1,
+	max: 2,
+	step: 0.1,
+	default: 0.5,
+};
+
+export const minimumSeparationOptions = {
+	variable: "minimum_separation",
+	inputType: "slider",
+	min: 0.001,
+	max: 0.25,
+	step: 0.001,
+	default: 0.05,
+};
+
+export const intersectionErrorOptions = {
+	variable: "error",
+	inputType: "slider",
+	min: 0.001,
+	max: 0.525,
+	step: 0.0025,
+	default: 0.02,
+};
+
+export const separationDiskDiameter = {
+	variable: "separation_disk_diameter",
+	inputType: "slider",
+	min: 2.5,
+	max: 25,
+	step: 0.1,
+	default: 5,
+};
+
+export const bezierTValueVariantOptions = {
+	variable: "TVariant",
+	inputType: "dropdown",
+	default: 0,
+	options: BEZIER_T_VALUE_VARIANTS,
+};
+
+export const subpathTValueVariantOptions = {
+	variable: "TVariant",
+	inputType: "dropdown",
+	default: 0,
+	options: SUBPATH_T_VALUE_VARIANTS,
+};
+
+export const joinOptions = {
+	variable: "join",
+	inputType: "dropdown",
+	default: 0,
+	options: JOIN_VARIANTS,
+};
+
+export const capOptions = {
+	variable: "cap",
+	inputType: "dropdown",
+	default: 0,
+	options: CAP_VARIANTS,
+};
